Store test db in memory
This ensures you can re-run tests without having to wipe the DB.
Consider changing this back if you're debugging, but otherwise this
is probably a good option.

If we want to test against a real db like mysql we will need a way to
wipe or recreate the database before tests are ran.