Prevent duplicate questioner entries
This is the only place we join sessions, so this is fine. However, it
may be nicer in the future to replace the Pivot we're using with one
that enforces unique constraints at the database level instead.
Right now, for example, if 2 join requests were sent at the same time,
this could cause 2 entries to be created still.