Support voting on questions
Introduce a small Vote model that keeps track of which users have
voted on which questions.
The double-vote prevention is not a database constraint so it may
be possible for 2 requests to come in quick enough and get around
this. I did not see any constraint support in Fluent's docs yet.